Added TTB Form Validation
Finished base validation for the Add TTB Form on the TTB field.  It
calls the CRUD service, which communicates with the database to check if
the value already exists.  If it does, the validator triggers ngMessages
to display an error message.  This is only checked when the user enters
and leaves the field.